These events depict daily patterns, weekly patterns, and unique actions.
Based on the specifications for our Market Street project, we created a
data model that simulates the flow of buses, cars, trucks, and
pedestrians along the Market Street corridor. The model contains many
patterns and events based on actual data that collected in late January
2014. The data set contains 160,000 data points describing two weeks of
pedestrian and vehicular circulation across four intersections on
Market Street, San Francisco in a 1 minute temporal resolution. You are
welcome to add third-party data to this data set
http://urbandatachallenge.org/urban-data-canvas/